Merged PR 271725: Upgrade to .net standard 2.0
Per the .net core team, we should be able to now build on .Net Standard 2.0, which is what we will eventually ship. This PR updates all of the csproj files to use .net standard 2.0. I've updated the build scripts as well.

**Important**
Your local version of the dotnet core sdk needs to be on preview1 for this to work. You will also need to upgrade to VS 15.3 so that the sdk is used instead of the sdk bundled with VS (this also fixes the rebuild issue with the dependency file)