Merged PR 281139: Edge BLE Module Part 1
Edge BLE Module
The changes here deal with creating a Azure Edge module with capabilities to connect to BLE devices to a local Edge Hub. This will eventually be delivered in a docker image and can be pulled down at will to execute within an Edge runtime.

The changes here deal with defining configuration permitting the module to connect to any BLE sensor device without the need to define a MAC address. Along with the configuration the changes contain a parser and interface implementation to print out the data parsed.

The logical structure of the configuration consists is hierarchical:
BLESensorsConfiguration  - Primary interface to parse the configuration and also serves as the container to retrieve one or more configured sensor hub config data.

SensorConfiguration - Logically this can be thought of as a sensor hub. This serves as a container for one or more physical sensor devices (temperature, humidity) etc. Additionally this has the necessary information for printing firmware, hardware rev etc.

DeviceConfiguration - This serves as the container and interface to obtaining the actual commands to communicate with a physical sensor.

Files
sensors.json contains the configuration to describe a TI Sensor Tag sensor hub and its devices that we are interested in communicating with. Additionally, each of these devices have commands associated with them that are defined as BT UUIDs that can be used to send to the sensor hub.

sensorsconfiguration.ts and deviceconfiguration.ts contains the interface and implementation to parse and read out the configuration.

main.ts At this time this is a simple program to assist with reading and displaying the configuration.

Other files are needed to bootstrap the project.