Merged PR 319990: Tail logs from stopped/paused container
When the MMA is reconciling modules, it removes (docker rm) containers that are stopped or paused and updates (docker create) them before attempting to restart them. If they were meant to be removed (as indicated by the desired properties), they would simply be removed (docker rm) and not restarted. In order to understand why a container stopped, the changes in this PR will print the last 25 logs from the container, before the container is removed. This will help us debug failed containers better.