Merged PR 305670: Edge BLE 2: Edge Module with noble
Edge BLE 2: Here is where the integration with noble and interface implementations to manage and access the BLE hardware has been implemented. The configuration is built with the temperature and optical sensor at this time.

The changes here deal with creating a Azure Edge module with capabilities to connect to BLE devices to a local Edge Hub. This will eventually be delivered in a docker image and can be pulled down at will to execute within an Edge runtime when its deps are ready.

The changes here deal with defining configuration permitting the module to connect to any BLE sensor device without the need to define a MAC address. Along with the configuration the changes contain a parser and interface implementation to print out the data parsed.

The logical structure of the configuration consists is hierarchical:
BLESensorsConfiguration - Primary interface to parse the configuration and also serves as the container to retrieve one or more configured sensor hub config data.

SensorConfiguration - Logically this can be thought of as a sensor hub. This serves as a container for one or more physical sensor devices (temperature, humidity) etc. Additionally this has the necessary information for printing firmware, hardware rev etc.

DeviceConfiguration - This serves as the container and interface to obtaining the actual commands to communicate with a physical sensor.

SensorManager: This is responsible for scanning and connecting to various sensor hubs and serves as the main sensor control interface query the different sensor hubs and their devices.

BLEEdgeModule - This is Edge module implementation for responsible for connecting with the IoT Hub, obtaining sensor hubs configuration via its twin and handling method calls to execute commands on the various devices. Additionally any telemetry data from the sensors is dispatched to the Edge Hub/Cloud via this module.

Main - Bootstrapper of the program