Merged PR 330905: Install docker for RS1 build machines.
Based on the review of the ARM template used to deploy VSTS machines, it would seem that Containers feature is already enabled. If that was the case, just installing docker should be sufficient to get going on RS1. I will still follow up separately to deploy a separate VM for RS3 bits.

Related work items: #1364016