Merged PR 362025: Remove incorrect assumption
When implementing the last set of changes to TwinConfigSource, I made the incorrect assumption that a module set could never be empty when TwinConfigSource::GetModuleSetAsync was called because it should have been populated by the OnConnectionStatusChanged handler. However, the module set can be empty, if the twin is empty! D'oh!