Merged PR 378209: Switch to AMQP between Edge hub and the cloud
The eventual goal is to switch to AMQP **over WebSockets**, but there's currently a .NET Core-specific bug in the C# device SDK which prevents it. So this change switches the backend connection in Edge hub from MQTT to AMQP (TCP). It also sets up multiplexing, aka connection pooling. We default to 10 connections in the pool, configurable via app settings.

Related work items: #1443160