Merged PR 378676: TwinManager to intercept twin requests
Sending this out to get feedback on the approach I'm taking on managing twin data. Currently the twinManager is just a pass through to the cloud and device. I was in the process of implementing the twin store, but since @<vpuranik@microsoft.com> is close to having one, I'll wait to integrate with that.