Merged PR 480992: Edge agent reports errors when they occur during reconciles
When we implemented reporting of edge agent state via its twin we had punted the work necessary to report failures that might occur at various phases during edge agent's lifetime. The phases of interest are:

1. Download of twin from IoT Hub
2. Deserialization of desired properties
3. Inspection of the runtime environment for fetching the _current_ state
4. Creation/update of module identities
5. Preparation of the execution plan during a reconcile
6. Execution of the plan during a reconcile

If things go wrong at any time during these 6 phases we want that reported in IoT Hub with as much information as is available while ensuring as little disruption in service availability as is possible. This PR tries to address this.

For the benefit of reviewers here's a potential sequence that you might want to consider that might help make sense of all the changes.

1. For phases 1 and 2, the code resides in `EdgeAgentConnection.cs`. It might make sense to begin the review there. The `EdgeAgentConnection` uses the IoT Hub Device SDK to fetch the twin asynchronously. It caches any exceptions that might get thrown in a member field and re-throws that when the call to `GetDeploymentConfigInfoAsync` arrives.
2. The `Agent` class in its `Reconcile` method calls the `IEnvironment` and `IConfigSource` objects to fetch the current state and the desired state. If those throw then that gets reported in the `Reconcile` method. This handles phase 2 and 3.
3. `Agent.Reconcile` works with the `IModuleIdentityLifecycleManager` object to create/update module identities. If this fails then `Agent.Reconcile` reports that error (phase 4).
4. Phases 5 and 6 are also handled from `Agent.Reconcile`.

Another aspect to consider is the behavior of the `FileBackupConfigSource` object which decorates `TwinConfigSource` by handling exceptions that might get thrown by `TwinConfigSource` by fetching the twin from a cached copy. In case of de-serialization errors we don't want the `FileBackupConfigSource` to quietly handle the error and hand out the cached copy. We instead want it reported. This PR handles this.

If a deployment has an empty modules list then that is an error. That is implemented as well.

Related work items: #1617129