Merged PR 498228: Allow setting additional tags for manifest images
We would like to keep a "moving" tag for the preview release so that the ui can provide a sensible default. For example we release `1.0.0-preview006` and tag with both `1.0.0-preview006` and `1.0-preview`. When `1.0.0-preview007` is released, we tag with `1.0.0-preview007` and `1.0-preview`.

This is similar to the `latest` tag but we can at least pin a minor version.

In order for this to happen, we need to allow additional tags on the manifest image. This PR implements that possibility.