Merged PR 506796: Rename docker images and python package in prep for release
Renames the four images we are shipping as follows:
* edge-agent --> `microsoft/azureiotedge-agent:<version>-<os>-<arch>`
* edge-hub --> `microsoft/azureiotedge-hub:<version>-<os>-<arch>`
* functions-binding --> `microsoft/azureiotedge-functions-binding:<version>-<os>-<arch>`
* temp sensor --> `microsoft/azureiotedge-simulated-temperature-sensor:<version>-<os>-<arch>`

This will prep us for releasing to Docker Hub. It also follows best practices for adding os/arch information (this info belongs in the tag).

Renames the python package to `azure-iot-edge-runtime-ctl`
Removes `buildDockerImage.sh`
Removes building `latest` images

This will require changes to the build after it is merged.