Merged PR 509548: Run edge hub as an unprivileged user
This PR includes changes for allowing Edge Hub to run with an unprivileged user account. The basic approach is to have the Dockerfile add a user (called _EdgeHubUser_) and then have `start.sh` use that user when launching the `dotnet` command. Being a non-root user means that we can no longer bind to TCP ports below 1024 which includes port 443. In order to allow this edge agent now adds the `NET_BIND_SERVICE` capability for the edge hub container when it is started up.