Merged PR 531225: Add logging handlers to avoid red error text for INFO, DEBUG, and WARNING in PowerShell
At present, python's logging module outputs to STDERR at all levels of logging. When running remote PowerShell (as is done when installing on IoT Core) or PowerShell ISE, this causes all logged text to be written to the terminal in red with a black background and exception information. This formatting signals to PowerShell users that an error has occurred even though this often is not the case. To avoid this, all log levels below ERROR must be written to STDOUT. This PR enables logging to STDOUT in such scenarios when iotedgectl is running on a Windows machine without affecting current Linux behavior.