Merged PR 542238: Added changes to make the restart policy as always for the Edge Agent
Added changes to make the restart policy as always for the Edge Agent.

This is needed so that when the Edge device is restarted, the Edge agent is automatically started and it can bring up the rest of the deployment. This is useful when operating the Edge as a gateway.

Note: iotedgectl stop will stop the Edge Agent and will need a subsequent iotedgectl start or restart to kick off the Edge deployment.