Merged PR 548099: Fixes for long running twin tests
I have been running twin operations (get desired properties and update reported properties) for long periods of time and looking at the logs to examine what is going on. This PR has some fixes and enhancements for issues observed via the logs. I only have anecdotal evidence that the "retry" (the ones in CloudEndpoint.cs) fixes in this PR fixed an issue with reported properties, but admit that I don't have full knowledge of intended design of the routing code to know if the fixes are correct. With this fix, the reported properties updates ran for more than 2 days, as opposed to it dying within 2 hours. If I'm missing something I would like to discuss and learn further.

The AtomicBoolean.cs change is to ensure we issue a memory barrier before checking the underlying value so that we don't read a stale value.

The logging change just makes the logging in routing a little better.