Merged PR 551564: Don't fail reported property update when twin is not cached
We had a corner case that would result in a reported property update failing if we couldn't cache or hadn't previously cached the twin. This could happen if we go offline before we were able to fetch the twin. These changes will prevent us from failing a downstream module's reported property update in that corner case.