Merged PR 565249: Log errors that happen very early in agent startup
Exceptions that bubble up to the agent's outermost exception handler were not being printed to stdout/stderr. So, for example, if the agent couldn't parse the device connection string, or if it tried to read an integer environment variable which the user set to all alpha chars, then it would terminate silently (`docker logs edgeAgent` returned nothing).

This fix does the following:
1. Any exceptions that land in the agent's outermost exception handler are printed to stderr.
2. The logger is started as soon as possible so we can use it to print more errors (it's true the outer exception handler will print them, but the logger lets you add a description and prints a more concise callstack).
3. Any exception thrown while [a] reading the configuration, [b] generating the agent connection string, or [c] building the agent application will print via the logger instead of the outer exception handler.