Merged PR 574296: Fix EdgeHub Proxy update and Subscription issues.
Issue1 - ConnectionManager.CreateCloudConnection call sets up a callback using the ConnectedDevice object which gets populated with the DeviceProxy. When the token expires, we call close on this DeviceProxy object.
However, when the identity is updated in between the 2 calls, a new ConnectedDevice object can get created, resulting in the close being called on the old deviceProxy, instead of the current device proxy.
This was fine earlier because we always used to create new device and cloud connection objects, so a new callback would get set up as well. Now, we always use the same CloudConnection object for a particular device, so the same callback is used for subsequent connections from the same device. As a result, the old ConnectedDevice object can get used in the callback for subsequent connections.

Fix1 - Instead of passing in the device object to the callback method directly, get it from the list of devices, so that we always get the current DeviceProxy object.

Issue2 - Subscriptions to EdgeHub were not getting persisted across re-connections
Fix2 - Persist connections across re-connections