Merged PR 572484: Enable SO_REUSEADDR for the Mqtt protocol head
If the Edge Hub doesn't cleanly shutdown, it can see an "address in use" error when restarting. Setting `SoReuseaddr` allows a listening socket to bound to the port if all of the other lingering sockets are in `TIME_WAIT`.

This should address the recent bug reports around:

```
(Inner Exception #0) Docker.DotNet.DockerApiException:
Docker API responded with status
code=InternalServerError,
response={
    "message":"driver failed programming external connectivity on endpoint edgeHub (1286aef053ce857655452f17670d4fe18c24f3db63a92a2728aae48d1e84e92b): Bind for 0.0.0.0:8883 failed: port is already allocated"
}
```