Merged PR 593654: Fix AsyncLock possible deadlock
There is a race condition between `ContinueWith` and `semaphore.WaitAsync`. Canceling the token between these two calls results in the lock never being released. We don't use a cancellation token in the `ContinueWith` because the task that it runs only creates an `IDisposible` that releases the lock, which we want to have run under all circumstances.

We share the same AsyncLock code as IoT Hub. They have discovered a deadlock in the AsyncLock class and have resolved it:

https://msazure.visualstudio.com/One/_git/Azure-IoT-Hub-Main/pullrequest/575854

We also should fix this in our code base.