Merged PR 601704: Handling Not Authorized Exception throw by Device Client on Parser.
Handling Not Authorized Exception throw by Device Client on Parser.

Bug Description:
Scenario: Edge Hub Configured with 2 modules (Temperature Sensor and Filter(Function)), running for 10+ hours.

Problem: EdgeHub Tries to renew token when it's near expire and it gets an exception from Device Client if the Token is expired.

This fix suppress and logs the unauthorized exception and disconnect the client so it can retry.