Merged PR 613261: Updating case from E2E Test Tag
Updating case from E2E Test Tag.
Details:
On our Unit Test Framework (XUnit) we are using Tags to identify tests. Our E2E Test tag (Attribute and Discover file that I changed), had a bug on the comment trait, where we expected the class to be E2E, but the class was actually E2e.
So, when I did the code review I changed the file cap using Visual Studio, but that didn't change the real file (so we have the class name case matching with the file name), so I'm fixing this by using git mv.