Merged PR 610869: IotHubReporter to clear twin reported properties if AgentState read fails
Modified IotHubReporter to clear twin reported properties if we can't
parse the reported properties into an Agent State.

If we can't parse the agent state, the Twin's Reported properties are not set correctly for edgeAgent to use, so we clear the reported properties and let the rest of the function run with a default Agent State object.