Merged PR 684087: Add WebSocket support to Edge Hub's MQTT protocol head
This change allows devices and modules to connect to Edge Hub via MQTT over WebSocket. It listens for WebSocket upgrade requests on the existing HTTP server (ASP.NET Core). When a request comes through, it sets up a DotNetty channel using all the same infrastructure as our existing MQTT channel. The rest of the pipeline is that same.

I also updated all our sample modules (tempSensor, tempFilter, Functions binding) to accept a `"ClientTransportType"` setting that instructs the SDK to connect via WebSocket. It accepts values that are stringified versions of the SDK's TransportType enum, e.g. `mqtt_websocket_only` or `mqtt_tcp_only` (default). The values are case-insensitive. To change the transport type in a deployed docker container, use the container's "create options":

```
{
  "Env": [
    "ClientTransportType=mqtt_websocket_only"
  ]
}
```

One important note: if you set up the callback in the C# device SDK to bypass server certificate validation (e.g., setting `MqttTransportSettings.RemoteCertificateValidationCallback(...) => true`), it _will not be honored_ for WebSocket because the underlying .NET WebSocket library doesn't support it. So if you use WebSocket (e.g., passing `TransportType.Mqtt_WebSocket_Only` to `DeviceClient.CreateFromConnectionString`) you must have the proper server certificate installed.

RE test coverage, this change _increases_ % covered for `Microsoft.Azure.Devices.Edge.Hub.Http`, but _decreases_ coverage for `Microsoft.Azure.Devices.Edge.Hub.Mqtt`, for a slight gain in coverage overall. The DotNetty channel is difficult to unit test because it's a subclass of DotNetty's `AbstractChannel`, which is rather large and complicated. I tried adding end-to-end coverage but that requires validating the server certificate, which our end-to-end tests are not set up to do. We'll need to look at how we can better test this code.

Related work items: #2160428, #2187535