Merged PR 682343: Abillity to authenticate a module with a certificate
This PR contains:
- Registration of a client cert validation callback with dotnetty. The callback presents the client cert, which hub saves for validation at a later stage in the MQTT protocol.
- Use of hub's CA cert chain to validate a module's cert. The module's cert must be signed by a CA in the hub's CA cert chain.
- Cert validation helper functions.
- Changes to hub dockerfile to make it run from Visual Studio.

For testing, I generated a cert for the tempSensor. This cert was signed by the Device CA that iotedgectl uses. I made changes to the Device Client to support modules connecting with certs. I verified that tempSensor was able to connect. I added another module in a similar way and verified that the tempSensor and the other module could exchange messages.

Still working on unit tests.