Merged PR 695933: Enable running deploy tests against pre-release images
Today the [Deploy] tests are hard-coded to run DockerHub images tagged "1.0-preview", so you can't run them against a pre-release version of the agent, hub, and temp-sensor images in another registry, e.g. edgerelease.azurecr.io.

This change fixes the tests to use the value of the existing registryAddress environment variable as the prefix for all image names. Additionally, the tests now look for a new environment variable, imageVersion, and use it as the tag for all images.

Related work items: #2197382, #2204685