Merged PR 698766: Add module traits
## edgelet-core

  - Trait definitions for `Module` and `ModuleRuntime` in `edgelet/edgelet-core/src/module.rs` - these traits define what an edge module should be and what API the runtime should provide. We would implement these in terms of Docker primitives and operations.
  - Place-holder definition of the `Edgelet` struct in `edgelet/edgelet-core/src/edgelet.rs` - code for helping bootstrap edge agent will live here

## docker-mri

This crate is meant to implement the `Module` and `ModuleRuntime` traits from `edgelet-core` in terms of Docker primitives and operations and includes:

  - The `Module` trait has an associated type called `Config` which is meant to capture implementation specific configuration information. This crate implements this via the `DockerConfig` struct in `edgelet/docker-mri/src/config.rs`.
  - The `DockerModule` struct implements the `Module` trait in `edgelet/docker-mri/src/module.rs`.
  - A place-holder `DockerModuleRuntime` struct lives in `edgelet/docker-mri/src/docker.rs` - this is not complete yet. The implementation of the `ModuleRuntime` trait will live here.

## edgelet-test-utils

This crate is meant to house utility code that is designed to aid unit testing. This commit includes the `JsonConnector` struct that can be used to provide mock implementations of the `hyper` client's transport in terms of serializing a given T. For an example of how this can be used, see the tests for the `DockerModule` struct.