Merged PR 711653: Convert xUnit [Deploy] test to an app
We have a single test in the [Deploy] category. It's a black-box end-to-end test that automates the "Quickstart" tutorial (e.g., https://docs.microsoft.com/en-us/azure/iot-edge/quickstart-linux). It's intended to be a smoke test or BVT. It was written as an xUnit test so that if we add more tests like it, we can easily discover and run them all. The downside is that xUnit is not built for end-to-end tests; for example, it is difficult to pass arguments to tests (currently we use environment variables). Another problem is that running the test with `dotnet test` requires the .NET Core SDK, which isn't supported on ARM32 platforms like Raspberry Pi. This limits our ability to quickly confirm basic functionality on any platform.

To fix these problems, I converted the test to a .NET Core app called "IotEdgeQuickstart". It can accept several command line arguments, or you can set environment variables. If you run it with no arguments but you have our Key Vault cert installed, it will run against our test IoT hub and daily docker images.

I've confirmed the app runs on Ubuntu, desktop Windows, and Raspbian Stretch Lite (on RPi3). Once this change is in, I can convert the Service Deployment build definition and Edge Release definition to use the new app instead of the old test. We can also convert the CD Release definition (Linux and Windows) to use this app instead of the release-*.sh and release-*.ps1 scripts (one TODO there is to remove the tempSensor module from the config when the app is finished).