Merged PR 723867: Add future versions of validation macros
This commit adds to the collection of validation macros, alternate versions which work in functions that return a `Box<Future<T, E>>` instead of a `Result<T, E>`. Every validation macro is now also available with an `f` prefix that can be used in such functions. `ensure_range!` for example is also available as `fensure_range!`.