Merged PR 729091: Handle Token update on CBS Token link
Overview:
Currently, EdgeHub Identity contains the identity of the client, and its credentials (connection string or token). This works for MQTT because the client connects once and provides the credentials at the same time.
In AMQP, we have a CBS link and tokens are sent over that link as the previous tokens are nearing expiry. To handle this, we would have to either -
- Create a new identity for the same client (which is odd and causes multiple identities for the same client floating around)
- Update the existing Identity (make it immutable)
- Or separate out the identity from the credentials.
This PR takes approach #3

Changes -
- Added a concept of client credentials, which contains both identity and the credentials.
- Cleaned up the Identity (which had a number of extraneous fields)
- Added a way for cloud connection to get the latest token from AMQP CBS link
- Updated tests to handle the changes