Merged PR 772583: Workaround service regression in EdgeHubConnectionTest
The service recently regressed the behavior of reported property patches. Until now, when you send an empty object {} the service would treat it as a no-op. Now it clears the corresponding property's value. This caused some Asserts in EdgeHubConnectionTest to fail. I've changed the Asserts so that they no longer fail. The fix should go in before the end of the month, and when it does these Asserts will start failing again. That will be our signal to restore them to their original values.

We could just comment them out, but then there's a risk they won't get re-enabled.