Merged PR 785386: Add an IPV4 only option to edgeHub
While dotnet core 2.1 preview 2 fixed several problems with IPv4 only operation, they were not enough to overcome the problems we saw with having an IPV4 only system.  I'd be interested if, after looking at this PR, if people think that the dotnet core should fix this.

We make a check if OS supports IPV6 and set the Listen/Bind address accordingly.