Merged PR 794480: Implement the KeyStore and Sign functions for edgelet-hsm
Implement the KeyStore and Sign functions from edgelet-core::crypto into an HSM crate.

In core crypto, modified the `Sign` trait to add an associated type.  This new associated type implements a new `Signature` trait.  This is so we don't have to copy data out of `TpmDigest`, which has a pointer to data allocated from the HSM library. `TpmDigest` only needs to implement `Signature::into_bytes()` for the signature to work.

Removed the static HSM objects in favor of tying the HSM subsystems to the KeyStore that is implementing it.