Merged PR 805998: Add tokio support for named pipes
This commit is the first in a series that will culminate with support for Windows named pipes in edgelet. This commit adds support for named pipe clients in Tokio. It wraps the named pipe implementation provided by the `mio-named-pipes` crate and adapts it to the requirements of Tokio. The key feature addition is implementation of the `AsyncRead` and `AsyncWrite` traits.