Merged PR 819548: Fix EdgeHub AMQP Tests
Issue - CI Build has been failing because of EdgeHub E2E Tests.
The root cause is that DeviceSDK does not use the passed in TransportSettings is some settings have been passed in previously (say by a different DeviceClient).
So the fix is to use a dummy DeviceClient before any of the tests run and pass in the desired TransportSettings.