Merged PR 828295: Edgelet certificate generation
Added functionality to create certificates (CA, server and client types) via OpenSSL API. This can be used to create self signed certificates or have certificates be generated using an issuer.

Key types supported are RSA and ECC both of which are supported in IoT Hub.

If no issuer is provided (self signed) the key type used is RSA. If an issuer certificate is provided, the key type to be generated is determined dynamically by reading the certificate.

Notes:
 - uMock tests to follow soon. Wanted to get this out for review and get feedback.
 - There is a known issue marked with a todo. There are a couple of options to resolve this and will be done later.