Merged PR 839658: Implement decrypt on workload HTTP endpoint
This change implements the HTTP handler for the decrypt API in iotedged. The tests use a stub of the Decrypt trait which simply reverses the input. This code is a mirror image of decrypt.rs. I spent some time trying to use a single, shared handler, but got bogged down in rust details. We can eliminate the duplication later.