Merged PR 866460: Refresh EdgeAgent config periodically
Currently EdgeAgent depends on DesiredPropertiesUpdate callback and Connection status changed callback to update its configuration. Both of those depend on Device SDK, and are known to not fire reliably. This could result in the device never receiving updates from the cloud. So this change is to periodically check with IoTHub for config updates.
The check frequency is configurable, and the check timer is reset if one of the 2 callbacks above have fired.