Merged PR 881453: Backup DPS provisioning result
Created a wrapper for a provisioning trait object that backs up credentials after provisioning completes. It uses the backup if provisioning fails. This is needed so that we can successfully provision the device if the DPS service cannot be reached.

We might use this wrapper for manual provisioning as well to detect if the settings have changed.