Merged PR 882059: Add third state to pid
Currently, the Pid type holds a value for a pid on Linux, or uses None for windows (since windows talks to docker via named pipes and named pipes can't supply the caller's pid). The equivalence logic allows pids to be compared on Linux, but access control is effectively disabled for named pipes (Pid::None == any pid). But this logic breaks the scenario where a system (like Unix domain sockets) *should* supply a pid but for some reason does not: the value is None, and access control mistakenly allows the operation.

To fix this, a third state, Pid::Any, is added. Any behaves like None used to. And None does not compare equal to any pid, ever.