Merged PR 888557: Edgelet Encrypt - Decrypt APIs
Added support for the following APIs in Edgelet along with unit and integration tests.
 - Create master key
 - Destroy master key
 - Encrypt
 - Decrypt

The encryption algorithm chosen is AAED GCM which lined up perfectly with Edge requirements for differentiating per module encrypt/decrypt requests plus it has error detection capabilities during decryption in case the cipher text was corrupted and/or a different module id was attempting to decrypt the data or the IV changed.

Yes the review in terms of number of files is large but most of the files are test related and the bulk of the implementation is in a few select files. Here is a summary of the files for review

- All the *.rs files are mostly in Edglet HSM and implement the edgelet core HSM traits for master key mgmt and encrypt/decrypt APIs. These also contain tests. Should be quick reviews.
- edge_enc_openssl_key.c contains the core implementation of enc-dec and the binary format for the resulting ciphertext. This is the largest file.
- edge_hsm_client_crypto.c, edge_hsm_client_store.c deal implement the framework pieces to invoke the enc/dec. Smallish changes.
- hsm_client_store.h, edge_sas_key.c, edge_pki_openssl.c, edge_hsm_key_interface.c contain implementation of a generic key destructor interface. edge_hsm_key_interface.c also has some minor cleanups. Pretty small and straight forward.
- The rest are umock tests related files.