Merged PR 895072: Fix arm package glibc dependency
The edgelet arm package requires glibc >= 2.23. This is purely a problem with how the package dependencies are listed, as the edgelet can support lower level versions of glibc.

This PR installs the proper cross build of glibc so that the dependency can be specified properly in the package.

Related work items: #2587502