Merged PR 903373: Quick fix for certificates generation on every Edglet restart
Implementation along with tests is not the complete fix rather this is a step in the direction where the create_certificate semantics are get or create if needed. What's missing is the validation checks for expired certificates. We do not ever want to accept or serve up expired certs. This checking will be done in a subsequent PR.

One other fix was included and minor cleanups.