Merged PR 909115: Increase dps timeout and reduce frequency of polling
The DPS timeout may be too short. Talking to the DPS team, we should wait at least 1 minute before giving up. I changed the timeout to 2 minutes. Originally I considered adding this timeout to the config file but I don't think its a knob we want to expose to customers.