Merged PR 915837: Update cloud identities only when container doesn't exist
Currently we get the edgeAgent identity from the cloud even if the container exists locally. We do this to get the generation id of the module. But we only use the generation id to create a new container. If the container already exists, we don't need the generation id. The current behavior poses a problem when we try to restart iotedged when offline. iotedged tries to connect to iothub to get the edgeAgent module identity and fails. This makes the daemon exit with an error.