Merged PR 923521: Windows environment variable read fix
When running the transparent gateway scenario on Windows, it was observed that env variable being set by IOTEDGED could not be read using stdlib API getenv. The result of this API was always NULL. Changes here create a new hsm util API to obtain a value of an environment variable. The implementation uses GetEnvironmentVariableA for Windows and getenv for *nix implementations.

Summary of changes
- hsm util API and tests
- Changed all code using getenv to use the new hsm_get_env API.