Merged PR 929764: Add ConnectionDeviceId and ConnectionModuleId to amqp outgoing messages
Outgoing messages for modules should contain ConnectionDeviceId and ConnectionModuleId properties indicating the source of the message. This already works for Mqtt. This PR adds support fo that on Amqp.
This also needed a corresponding change on the SDK side, here is the PR for that - https://github.com/Azure/azure-iot-sdk-csharp/pull/537