Merged PR 931371: Make Makefile all build iotedged in debug
This allows running simply `make` to produce a runnable binary. The current version will build a release version, but when run, it can't find the libiothsm.so.

This Makefile was largely made for the packaging, but we should be producing something runnable when calling `make`.

Addresses github issue: https://github.com/Azure/iotedge/issues/3