Enable iotedged identity certificates endpoint (#497)
The changes here enable modules to request their identity (client) certificates. With this capability, modules built as microservices and their clients can authenticate themselves in a generic manner using standard secure communication protocols built over TLS!

Here are the salient properties of the certificate.

These certificates are rooted using the Edge device CA.
Note: The CA certificate is obtained via the trust bundle API.
Each of these certificates will have their identifier stamped in the certificate as SAN URI entry which can be used for identification. The following is the URI format:
azureiot://{hub}/devices/{did}/module/{mid}
The CN field of the subject will contain the module id and is not configurable.
The certificates are short lived for a duration of two hours.