iotedged should return 304 not 500 (#560)
We recently began seeing a problem where iotedged couldn't restart if it didn't give edgeAgent time to gracefully shutdown all modules before it went down. This is a problem on Windows where iotedged as a Windows service is terminated by the OS after five seconds.

The problem was that some error translation logic in our iotedged management endpoint was looking in the wrong part of the error coming from our core, and therefore not finding what it was expecting (304 not modified) and returning 500 (internal server error). The fix was to update the translation logic to look in the right place.

I also added a unit test for the 304 error specifically, and updated the existing ones.