Use HTTPS proxy in Edge Hub (#697)
In Linux, .NET Core recognizes and honors the `https_proxy` environment variable when it is present in a module like Edge Agent or Edge Hub. The behavior is different in Windows, however, where the default WinInet proxy settings are used (as set via the Control Panel, or Internet Explorer) and `https_proxy` is ignored. This is especially a problem in RS5 nanoserver containers, which don't even expose the WinInet proxy settings.

The fix is to look for the environment variable ourselves, create a `WebProxy` object, and attach it to the `ITransportSettings` object that we pass into the SDK's `ModuleClient`. We technically only have to do this for Windows, but we'll do it for all platforms, for consistency.

This change updates Edge Hub. Edge Agent was updated previously (see #687). Note that for Edge Hub, we have to pass the proxy info into our own `System.Net.Http.HttpClient`-based `DeviceScopeApiClient` as well as `ModuleClient`.