Don't lower-case the keys when parsing config.yaml (#1038)
`struct DockerConfig` is set up so that its serde Deserializer expects a field
named `createOptions`. However the config crate's YAML parser normalizes
all keys to lower case. Even if the user sets `agent.config.createOptions` in
`config.yaml`, the serde Deserializer will see a key named `createoptions` and
discard it as unknown. The actual `create_options` field is then set to an
empty default.

This prevents users from having custom create options for the Edge Agent
container.

With this change, the code no longer use the `config::File` source to parse
`config.yaml`. Instead, it uses a new `YamlFileSource` that also impls
`config::Source`. This alternative source has similar code to `config::File`
but does not lower-case the field names.

Since the `config::Environment` source still lower-cases the keys that it
parses from environment variables, it does mean that it will not be able to
override keys in the config.yaml that are of mixed-case. For example,
`IOTEDGE_FOO` will override `foo` and not `Foo`. In practice,
all the top-level keys that could be easily overridden are completely
lower-case, so it is not likely to break any customers.

Fixes #1036