Add structured logging to end-to-end test library (#1153)
This change uses Serilog to add structured logging to the temp sensor test. There's stuff happening in the test itself that will eventually get moved into the common library, but for now the temp sensor test is the only consumer so there isn't the need for an abstraction just yet.

Information-level logs are always sent to the console. If you pass the `--verbose` argument, it will also send verbose-level logs to the console. If you pass a `--logfile <path>` argument, it will also dump all logs (information + verbose) to a file. The idea here is to keep the console visually uncluttered (which you can override with `--verbose`), but create a logfile with all the gory details, which in our CI system will be added to the build's artifacts.

I was spending too much time on the formatting/display of elapsed times for information-level logs. I'm not even sure I'll keep them since the logs already have a timestamp, so I'll leave the elapsed stuff alone for a while, and reconsider it when we add more tests and flesh out the common library. But for now, it's there and it works.