Update package.ps1 (#1270)
Previously we defined $Version for the Windows 10 kits version, however, the same $Version was already defined as arguments to the function Function New-Package([string] $Name, [string] $Version), later on when pkggen is called, the mistakenly updated $Version is passed to pkggen as the version of the package. so we have 10.0.17763.0 as the ARM32 cab version, while we should have something like 0.1.19149.2. This change fixes the var namings so we don't overwrite the argument $Version, so ARM cab can have a good increasing version.