Update issue template to ask for `iotedge check` output (#1287)
Also provide a placeholder for docker version and IoT Edge component logs
that are wrapped in a code block. For the IoT Edge component logs,
their code blocks are also wrapped in `<details>` so that they're collapsed
by default.

Also mention the `-H` CLI parameter for Moby on Windows for
the `docker version` command.