Support running the Windows installer script in 32-bit PS host on 64-bit Windows (#1293)
Currently the installer script just uses the `ProgramFiles` env var.
But in the case of 32-bit PS host on 64-bit Windows, this env var
points to `C:\Program Files (x86)` instead of `C:\Program Files`.

With this change the script uses the `ProgramW6432` env var if it exists,
since that always points to `C:\Program Files` regardless of
the process bitness. However that env var does not exist on
an actual 32-bit host like IoT Core ARM32, so the script still falls back
to the `ProgramFiles` env var.