Add module-to-module direct method end-to-end tests (#1260)
The change adds the module-to-module direct method test to our growing suite of end-to-end tests that will replace IotEdgeQuickstart. The test is parameterized on the module-to-edge protocol, so there are four tests in all (AMQP, MQTT, AMQP-WS & MQTT-WS).

Now that there's more than one test in the suite, I've updated the infrastructure so that we run setup (i.e., create a new device identity and install IoT Edge daemon on the test device) and teardown once for the _suite_, instead of per test. To do that I replaced the console app with a regular test framework. Our existing unit/integration tests use Xunit, but I chose NUnit for the end-to-end tests because it can output log messages to the console _during_ the test run, instead of just at the end. Since the tests run for several minutes each and we're impatient, it's nice to be able to look at the console (or build dashboard) and see that things are moving along.

I've refactored a few things in the common library as needed for this change, but there will certainly be more as we add more tests. I still think each test should be shorter and easier to read, but we need to add a wider range of tests (like e.g. the transparent gateway test) before I know how best to refactor them.