Allow tests to set OptimizeForPerformance on edge hub (#1368)
This change helps the new end-to-end tests run on Raspberry Pi. If you don't set OptimizeForPerformance in the test settings (context.json), or if you set it to `true`, then the tests will _not_ set the environment variable in edge hub. If you set it to `false`, then the tests will set the environment variable OptimizeForPerformance=False in edge hub.