Remove unused script (#1400)
Given this notice about the potential for keyserver attacks:
https://gist.github.com/rjhansen/67ab921ffb4084c865b3618d6955275f

...and given that the one thing in our repo that uses a keyserver (`apt-key --recv-keys ...`) is an old script no one's using anymore, I'm removing the script.

The other option would be to update the part of the script that installs the Azure CLI. The steps in the latest docs use `curl` to grab https://packages.microsoft.com/keys/microsoft.asc--a method which shouldn't be vulnerable to the same attack. But I checked our CI builds and can't find any reference to this script, so it doesn't seem worth the effort.