Remove support for all config formats except for YAML and JSON (#1422)
62f3d44f9da239f5e6ef4a5637df14d93c8a5fe3 updated serde-hjson to 0.8.2,
which broke the Windows ARM32 build since we patch serde-hjson 0.8.1
for that target.

Furthermore, a39543ce1924e7d16b27c43d66bbbc5eb0a7ca29 then updated config
to 0.9.3 which requires serde-hjson ^0.8.2, so serde-hjson cannot be
downgraded any more.

Since we don't use any format other than YAML for release and JSON for tests,
this change only enables the features for those formats. This drops the
serde-hjson dependency entirely.

Also, this change removes the patch for the cmake crate.
9001b8e86b4e7323e970542760b13016c5ae5ace bumped cmake to 0.1.40
so the existing patch for 0.1.30 was not being applied. However the build
succeeds with 0.1.40 anyway, so it looks like the patch isn't necessary
anyway.