Add transparent gateway end-to-end tests (#1454)
This change adds the transparent gateway tests for Windows and Linux. It also moves common code among all tests into the library, so that the #LOC per test is much less. If you look at the code in the `[Test]` methods (see Device.cs and Module.cs), it should be fairly easy to read and understand what is happening. The goal is to make these tests easier to understand, maintain, and expand.