Remove Edge identity certificate auto generation (#1495)
Changes here remove the ability to generate a device identity certificate from the gateway CA certificates. This was removed from the ESD but not cleaned up in the HSM lib.

Most of the changes are in the tests to adjust for this modification. Additionally augmented the unit tests for the private key sign interface which were not implemented.