Update end-to-end build YAML for new tests (#1727)
Two new end-to-end tests have been added recently, but they both fail in automated builds because they rely on new context.json parameters and secrets that haven't been added to the YAML build. This change fixes that.

Note this change also makes the context.json parameter "caCertScriptPath" optional because it's only required for the TransparentGateway tests.