Support dynamic reprovisioning of devices in the external provisioning mode (#1842)
Support dynamic reprovisioning of devices.

If the dynamic reprovisioning feature is enabled, iotedged will simply shut down when a possible reprovisioning event is signaled by EdgeAgent (through a call to a new API on the mgmt endpoint /device/reprovision). The reason behind this is so as to allow a subsequent restart of the iotedged to trigger the device provisioning flow again (based on the appropriate provisioning mode) and if the device is provisioned on another hub, iotedged will be configured using the new provisioning information of the device.

For the external provisioning mode specifically, the iotedged will also notify the external endpoint about a 'possible' device reprovisioning event by calling a new /device/reprovision of the external endpoint.

On receiving such a notification, the external endpoint can check whether the device has indeed been de-provisioned from it's original IoT Hub and provisioned on another IoT Hub instead.
The external endpoint can then take appropriate actions to reconfigure the device such as cleaning up any cached local state and restarting IoT Edge.

More details about the feature can be found in the updates made to the ExternalProvisioning.md file.