Support module priority in IoT Edge (#2186)
With this change, a new property ‘Priority’ will be added to the deployment configuration of each module. This value will be used by Edge Agent to determine the order in which all other modules are to be created.

The value of ‘Priority’ will be positive and zero-based with 0 being the highest priority value. A higher numeric ‘Priority’ value would indicate a lower priority assigned to that module. The maximum value of this property will be UInt32.MaxValue (4294967295).

Modules that possess the same priority will be created at the same time and will have no deterministic startup order imposed amongst themselves. Modules with higher priority values will be created and started first and once Edge Agent has made an attempt to create and start them (only the modules with a desired state of ‘Running’ will be started), it will proceed with the creation (and optionally startup) of other lower priority modules in order.

Edge Agent will not support the ‘Priority’ property and will by default have the highest priority of zero (0).

Modules that do not have a ‘Priority’ value specified will be started in a non-deterministic order (as is the case today) and will be assigned a priority of UInt32.MaxValue which indicates the lowest possible priority. Edge Agent will trigger the creation of these modules after all modules with a higher priority have been created.

Module boot ordering will initially not be supported for the Kubernetes execution mode of IoT Edge.