Eliminate redundant deployments from end-to-end tests (#2442)
We recently split up end-to-end test deployments into two configs: one for just the system modules and another for all modules. The problem is that our tests will double-deploy even when the config _only_ contains system modules.

This change only produces/deploys the second config if it differs from the first. Tests that only deploy $edgeAgent and $edgeHub won't do it twice.