Configure Kodiak as a tool that automatically merge PRs (#2441)
In order, to prevent semantic merge conflicts having only GitHub protection policy is not sufficient. We need an automated tool that does both:
- automatically update the PR branch with the latest changes from master
- maintain a queue merge PRs into masters to block conflict changes

1. Configure Kodiak for the repo
2. Update check-ins to prevent unnecessary jobs to run
3. Support auto-merge with "ready-to-merge" label

When the CI tool starts it checks out a PR branch and makes a merge commit. This merge commit contains all changes made in PR brach. To prevent unnecessary jobs to run additional job task gets a list of all files that changed in the commit and identifies should the agent run all other jobs in the pipeline. The rest of the jobs in the pipeline wait until this check is finished, and either starts or skips execution. This way even Kodiak forces to run check-ins it will not cause unnecessary work to run.