Add .NET code coverage to the checkin build (#2393)
The change enables code coverage analysis for our .NET Core code in the checkin build, and (similar to our iotedged checkin build) fails the build if % line coverage is below a certain threshold. Current coverage is measuring at 61% so I set the threshold to 60%.

Note this change relies on the built-in code coverage functionality that has recently been added in .NET Core. Downsides are: it's currently only available on Windows, and it produces a .coverage file which has to be opened in Visual Studio. Upsides are: this is the "officially supported" option for Azure DevOps, so while support is minimal now, future improvements to analytics in DevOps would likely come for free. Also, because this option works out-of-the-box in the dotnet tooling, I was able to remove a **bunch** of gunk from our project files.

The other option would be to use the 3rd-party [Coverlet](https://github.com/tonerdo/coverlet) tool, which would allow us to run on Linux (might shave a couple minutes off job time) and outputs in Cobertura format just like our Rust code coverage. More details [here](https://docs.microsoft.com/en-us/azure/devops/pipelines/ecosystems/dotnet-core?view=azure-devops#collect-code-coverage-metrics-with-coverlet). So we'd have a more consistent experience between languages. But we'd have to add back a bunch of project references in our test projects for arguably minimal benefit since this change seems to give us everything we need, at least for now.

This change also adds the "Unit" category to 132 unit tests which were not categorized.