`Iotedge check` - Time drift overreporting fix (#2487)
This fixes an overreporting issue in the ContainerLocalTime check.

The previous implementation queried host time first and container time second, in effect factoring container startup time into the difference between the two. This is fixed by changing the order of the queries - container first, host second.

This problem was mentioned in #2130 by @JeeWeetje and the fix alluded to in this [comment](https://github.com/Azure/iotedge/issues/2130#issuecomment-566322950) by @kkdawkins.