Make logs upload method since take relative times (#2499)
This changes the since parameter for built in log upload to accept relative times: https://github.com/Azure/iotedge/blob/master/doc/built-in-logs-pull.md

### Edgelet
It moves the parse_since function in support bundle to a shared func in core and uses it for logs.

### EdgeAgent
Chances since from an int to a string. This is directly passed to edgelet.
Also changes management yaml