Updating tolerance period for DirectMethod connectivity tests (#2489)
Updating tolerance period for DM connectivity tests. We are getting a few failures, but our tolerance period is extremely low. This will bump up the tolerance period. If we see no failures, then we know that those issues were happening within the tolerance periods and were likely actually successful results.

(Tolerance period is defined as the period after we flip the NetworkController to disabled or enabled where we find faults to be acceptable. This is because of a delay in edgeHub realizing it is offline)