Fixing twin tests for connectivity (#2631)
We are seeing a bug in Connectivity Twin Tests. 

When we start the tests, we do a cleanup - where we grab the current twin, set its properties to null, and send an UpdateReportedProperties call. Due to how the SDK works, this will automatically subscribe us for receiving DesiredPropertiesUpdates (DPU). But then we wait for 2 minutes due to the testStartDelay. After that, we set the callback for receiving DPUs.

The issue is that if we receive a DPU during the testStartDelay, the callback won't be set, and we will do nothing. 

So this PR moves the callback up before the testStartDelay, so that we are ready to receive even during the waiting period. 

It also removes the cleanup because it is unnecessary now that we are using a fresh device every time.