Changing case of transportType for C2D connectivity tests to match deployment files (#2667)
Typo in C2D connectivity test files: wrong casing.

In the deployment files, we specify that the case is "transportType," but in the Settings file of the C2DTester module, we match to "TransportType." Therefore, the actual transportType setting won't get picked up by the C2DTesterModule, and it will always pick the default. 

This change fixes the casing. Now, the correct transportType will get picked up.