Fixing C2D sender for connectivity tests (#2659)
When c2d sender throws an exception it crashes. This was the original intent, since the sender shouldn't fail. That assumption turned out to be wrong, though.

it will fail if it exceeds 50 messages that are in the cloud queue. So if edgeHub is offline and 50 messages build up, we should just keep trying to send. Eventually the queue will be drained, and we'll be able to continue sending.