Install the latest linux packages in the end-to-end tests (#2726)
On Linux, the end-to-end test framework was following outdated instructions to configure the repository for apt packages. This meant it didn't pick up the new 1.0.9 release, which was only pushed to the multiarch repository. I updated the code to grab the multiarch configuration, so the test framework should be able to install the latest version of edge.