DirectMethodSender: Don't report internal server error for transient exceptions (#2805)
Every time the DirectMethodReceiver module's upstream connection is down (i.e. startup / edgehub restart), then DirectMethodSender will receive a DeviceNotFound exception when sending a direct method. 

Currently, we treat this as a internal server error when reporting results to the TestAnalyzer / TRC. Instead, we should return 404 and implement logic in the TestAnalyzer / TRC to treat this as a transient error.