Longhaul: Don't fail tests for status code 0 (#2790)
Longhaul direct method tests will often fail with status code 0 returned by edgehub to the sender module.

I talked to Priyanka on the SDK team who informed that the feature enabling the fix for SDK issue [1026](https://github.com/Azure/azure-iot-sdk-csharp/issues/1026) is still on the backlog. Until this is complete, we should not fail the tests for status code 0.