Fix host and docker cpu stats (#2879)
This pr does several things to fix cpu stats for docker and host.
* For host cpu metrics, the measurement object was added as a member of the runtime instead of being newly constructed each time. This allows it to measure cpu impact between measurements instead of only measuring the instant it is polled.
* For docker metrics, previous use is stored in a dictionary. This also allows the cpu use to reflect the entirety of the time between measurements instead of the current use when the measurement occurs.
* This also fixes a math error when converting from ms to ns.

It also changes memory stats to more closely reflect docker stats. Before it reported to reserved RAM, regardless of how much was in use. The `docker stats` command subtracts the memory cache from this value, so the metrics were modified to do the same.